requirements_streamlit.txt
# Nokia 5G Fronthaul Optimization - Streamlit Application Requirements

# Core Streamlit
streamlit>=1.28.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Image handling
Pillow>=10.0.0

# Network graph visualization (optional, for future enhancements)
networkx>=3.1
